+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Build for Raspberry Pi architecture (`alpine.3.16-arm64`)
+- [**BC**] Remove `Google Sheets` integration
+- Add Command for running a web server with sensors data (**currently dummy data**)
+- Add logger
+- Fix retry on unauthorized action
 
 ## 1.2.0 - 2022-09-12
 - Add commands